backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.27.0.post1
numpy==1.26.4
pillow==10.2.0
tensorflow==2.20.0
python-multipart==0.0.9
